fix(ansi): Fixed bug with unterminated sequences.
Fixed a bug where split packets resulting in unterminated ANSI
sequences in the middle of a line are filtered into incorrect
NSAttributedStrings for display. Changed to delay the parse
until all ANSI sequences in the buffer are complete.

This change may potentially introduce longer delays under high
network latency conditions, but avoids misrendered text. An
example of the error can be demonstrated using the code here:
https://github.com/hotspare/TelnetTestWidget commit 4b58e03.